aggiunto middleware per l upload di foto backend e interfaccia grafica per caricare una foto in un formato standard

+++ backend/controllers/authController.js
@@ -66,9 +66,11 @@ export class AuthController {
 
     static async resetDatabase() {
 
-        blacklistedTokens.clear();
-        return await database.sync({ force: true });
-    
+        if (process.env.DEV_MODE === "true") {
+            blacklistedTokens.clear();
+            return await database.sync({ force: true });
+        } else
+            throw new Error("resetDatabase can only be called in DEV_MODE.");
     }
 
 }
